Weather_Backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
babel==2.16.0
bcc==0.1.10
billiard==4.2.1
blinker==1.8.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
cron-descriptor==1.4.5
cryptography==43.0.1
defer==1.0.4
distro==1.9.0
distro-info==1.0
Django==5.1.1
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
httplib2==0.22.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.2
launchpadlib==2.0.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
netaddr==1.3.0
netifaces==0.11.0
numpy>=1.18,<2.0
oauthlib==3.2.2
olefile==0.47
packaging==24.1
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pycairo==1.27.0
pycparser==2.22
Pygments==2.18.0
PyGObject==3.50.0
PyJWT==2.9.0
pyparsing==3.1.4
pyrsistent==0.20.0
pyserial==3.5
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-debian==0.1.49
python-dotenv==1.0.1
pytz==2024.2
pyxdg==0.28
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
traitlets==5.14.3
traittypes==0.2.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wadllib==1.3.9
wcwidth==0.2.13
wheel==0.44.0
xdg==6.0.0
timezonefinder==6.2.0
redis==4.5.1
social-auth-app-django


